flask_app/requirements.txt
Flask==2.3.3
requests==2.31.0
Werkzeug==2.3.7
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
dnspython==2.4.2